+++ tests/integ/modin/groupby/test_groupby_idxmax_idxmin.py
@@ -8,12 +8,12 @@ import pytest
 
 import snowflake.snowpark.modin.plugin  # noqa: F401
 from tests.integ.modin.groupby.conftest import multiindex_data
-from tests.integ.modin.sql_counter import sql_count_checker
 from tests.integ.modin.utils import (
     assert_frame_equal,
     create_test_dfs,
     eval_snowpark_pandas_result,
 )
+from tests.integ.utils.sql_counter import sql_count_checker
 from tests.utils import running_on_public_ci