+++ tests/integ/modin/series/test_add_suffix.py
@@ -8,8 +8,8 @@ import pandas as native_pd
 import pytest
 
 import snowflake.snowpark.modin.plugin  # noqa: F401
-from tests.integ.modin.sql_counter import sql_count_checker
 from tests.integ.modin.utils import create_test_series, eval_snowpark_pandas_result
+from tests.integ.utils.sql_counter import sql_count_checker
 
 TEST_ADD_SUFFIX_DATA = [
     "_suffix",